Cleaned up all warnings and renamed Physx_Agent to PhysxAgent for consistency.

+++ HardLight/src/GUI.cpp
@@ -29,7 +29,7 @@ void GUI::loadMenu(const char * menuPath, PxPhysics* physics) {
 	string imagePath;
 	Image image;
 	Button button;
-	double x, y;
+	float x, y;
 
 	loader = new INIReader(menuPath);
 	if (loader->ParseError() < 0) {
@@ -44,8 +44,8 @@ void GUI::loadMenu(const char * menuPath, PxPhysics* physics) {
 	for(int i = 0; i < numOfImages; i++) {
 
 		imagePath = loader->Get("images", string("image") + to_string(i+1), "");
-		x = clamp(loader->GetReal("images", string("x") + to_string(i+1), 0.0f), (double)0.0f, (double)1.0f);
-		y = clamp(loader->GetReal("images", string("y") + to_string(i+1), 0.0f), (double)0.0f, (double)1.0f);
+		x = clamp((float)loader->GetReal("images", string("x") + to_string(i+1), 0.0), 0.0f, 1.0f);
+		y = clamp((float)loader->GetReal("images", string("y") + to_string(i+1), 0.0), 0.0f, 1.0f);
 
 		image = Image(imagePath.c_str());
 		image.setPos(vec2(x, y));
